updated python 2 to 3 and update to QT6
* pytho  2to3

* WTO3

* 2to3

* updated qt6

* example mcnp input

* qt 6

* python path

* print statement check

* remove python2 comma

* pep

* line carriage return change

* python path edits

* mouse wheel event

* povray edit

* temp folder

* syntax

* generic plane eq

* reg exp update

* undefined int

* mouse zoom

* file path

* signal slot fix

* bb updates

* tidy up print

+++ source/MCNPXVisualizer.h
@@ -208,7 +208,7 @@ class MCNPXVisualizer : public QMainWindow
 		PythonBinder* _pythonBinder;
 
 		// MATERIALS
-		int _currentColorIndex;
+        size_t _currentColorIndex;
 		std::vector<QColor> _colorMap;
 		std::map<QString, Material*> _standardMaterials;
 		std::map<int, Material*> _savedMaterials;
